Add category restrictions to roles
We currently have the concept of category restrictions for users, which
restricts which categories a user can operate in. However this can be a
bit cumbersome when you want to restrict the categories for all users in
a role.

This change adds category restrictions to roles, along with an admin
interface for editing them.