[Bromley] Category restriction for role restricts report creation
Previously, category restrictions for roles worked the same as they do
for users, which is that they just restrict when the inspect form
appears.

Bromley want to also restrict which categories a report can be created
in when a role has category restrictions.

For now I've made this change Bromley-specific, but shouldn't be too
tricky to make this generic if we decide to go that way.